move myPlanet reports component to myplanet sub directory
