+++ src/app/manager-dashboard/reports/myplanet/logs-myplanet.component.scss
@@ -1,4 +1,4 @@
-@import '../../variables';
+@import '../../../variables';
 
 $screen-md: 1300px;
